requirements.txt
# Google Cloud Libraries
google==3.0.0
google-api-core==2.8.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
db-dtypes==1.3.1

# Application framework
fastapi==0.112.2
uvicorn==0.30.6
nest_asyncio==1.6.0

# Environment and Configuration Management
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.4

# Timezone Support
pytz

# Data Visualization Libraries
matplotlib==3.9.2
matplotlib-inline==0.1.6
seaborn==0.12.2

# Numerical and Scientific Computing
numpy==1.26.4
pandas==2.2.2
scipy==1.14.1
xgboost==2.1.2
joblib==1.4.2
openpyxl==3.1.5

# Data providers
ccxt==4.3.94
polygon-api-client==1.14.2

# Web3
